+++ school-service/src/main/java/com/ushirikeduc/schools/requests/DirectorResponse.java
@@ -4,7 +4,7 @@ import com.ushirikeduc.schools.model.Address;
 
 public record DirectorResponse(
         String name ,
-        int schoolID ,
+        long schoolID ,
         Address address
 ) {
 }
